+++ src/controls.rs
@@ -1,5 +1,6 @@
 // Declare the control handler sub-modules
 pub(crate) mod button_handler;
+pub(crate) mod combobox_handler;
 pub(crate) mod dialog_handler;
 pub(crate) mod input_handler;
 pub(crate) mod label_handler;
@@ -7,6 +8,7 @@ pub(crate) mod menu_handler;
 pub(crate) mod paint_router;
 pub(crate) mod panel_handler;
 pub(crate) mod progress_handler;
+pub(crate) mod radiobutton_handler;
 pub(crate) mod richedit_handler;
 pub(crate) mod splitter_handler;
 pub(crate) mod styling_handler;